+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 - Load from template method now parses the report through the Report parser.
 - Plain strings in the model report method list return as a string rather than nil
+- Arrays starting with a string containing only a hash attempt to use the second value as target
 
 ## [0.1.0] - 2022-10-06
 - Initial release