+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.3] 2022-11-28
+### Fixed
+- Target parsed by ValueParser.parse_and_execute method no longer is duplicated
+  - This fixes a bug where active record relations would not properly function
+  - ValueParser.parse has now a third parameter, a boolean on whether to duplicate
+
 ## [0.3.2] 2022-11-24
 ### Fixed
 - Reference to schema file